Continuing fixes for refactor. Wish we had compile-time checks
